config(configs): tsconfig/bitest configs move to project root
