+++ packages/@aglabo-actions/tool-installer/package.json
@@ -34,8 +34,8 @@
     "lint:types": "pnpm exec eslint --config ./configs/eslint.config.typed.js --cache --cache-location .cache/eslint-cache/ ",
     "lint:secrets": "secretlint --maskSecrets --secretlintignore .gitignore **/*",
     "test": "pnpm run test:develop && pnpm run test:ci",
-    "test:ci": "pnpm exec vitest run --config ./configs/vitest.config.ci.ts",
-    "test:develop": "pnpm exec vitest run --config ./configs/vitest.config.unit.ts ",
+    "test:ci": "pnpm exec vitest run --config ./vitest.config.ci.ts",
+    "test:develop": "pnpm exec vitest run --config ./vitest.config.unit.ts ",
     "test:watch": "pnpm exec vitest --watch",
     "sync:configs": "bash ../../../scripts/sync_configs.sh . "
   },
@@ -48,5 +48,8 @@
     "@actions/io": "^1.1.3",
     "@actions/tool-cache": "^2.0.2",
     "command-exists": "^1.2.9"
+  },
+  "devDependencies": {
+    "@types/command-exists": "^1.2.3"
   }
 }
