+++ packages/@aglabo-actions/tool-installer/src/helper/__tests__/getExistEget.spec.ts
@@ -8,13 +8,16 @@
 
 // src/installer/__tests__/ensureEget.spec.ts
 import commandExists from 'command-exists';
+import type { MockedFunction } from 'vitest';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { getExistEget } from '../egetInstallHelper';
 
-// command-exists を default モックとして差し替え
-vi.mock('command-exists', () => ({
-  default: vi.fn(),
-}));
+// 元の関数（モック対象）があるモジュール
+vi.mock('command-exists', () => {
+  return {
+    default: vi.fn(),
+  };
+});
 
 describe('isEgetAvailable', () => {
   beforeEach(() => {
@@ -22,7 +25,10 @@ describe('isEgetAvailable', () => {
   });
 
   it('eget が存在する場合 true', async () => {
-    vi.mocked(commandExists).mockResolvedValue('.tools/bin/eget');
+    const mockedExists = commandExists as unknown as MockedFunction<(cmd: string) => Promise<string>>;
+
+    mockedExists.mockResolvedValue('.tools/bin/eget');
+
     const result = await getExistEget();
     expect(result).toBe('.tools/bin/eget');
   });
