+++ packages/@aglabo-actions/tool-installer/tsconfig.json
@@ -1,21 +1,21 @@
 {
   "@(#)": "tsconfig.json: for @aglabo-actions/tool-installer",
-  "extends": "../../../../shared/configs/tsconfig.base.json",
+  "extends": "../../../shared/configs/tsconfig.base.json",
   // directory definition
   "compilerOptions": {
     // in/out dir
-    "rootDir": "../",
+    "rootDir": "./",
     "outDir": "lib",
     // aliases
     "baseUrl": "./",
     "paths": {
       // ローカルのプラグイン、ルール
       "@/*": [
-        "../src/*"
+        "src/*"
       ],
       // Monorepo 共有用
       "@shared/*": [
-        "../../../../shared/*"
+        "../../../shared/*"
       ]
     },
     // incremental build info
