+++ packages/@aglabo-actions/tool-installer/vitest.config.ci.ts
@@ -20,7 +20,7 @@ import { fileURLToPath } from 'url';
 import { defineConfig } from 'vitest/config';
 
 // user common config
-import baseConfig from '../../../../shared/configs/vitest.config.base';
+import baseConfig from '../../../shared/configs/vitest.config.base';
 
 // constants
 const __dirname = dirname(fileURLToPath(import.meta.url));
@@ -34,6 +34,7 @@ export default defineConfig({
   resolve: {
     alias: {
       '@': path.resolve(__dirname, 'src'),
+      '@shared': path.resolve(__dirname, '../../../shared'),
     },
   },
 });
